Se crea un subpaquete por cada entidad para agrupar sus vistas por dominio.
